financial_management/requirements.txt
Django==3.0.4
psycopg2-binary==2.8.3
gunicorn==19.9.0
django-datatable-view==0.9.0
django-crispy-forms==1.9.0
django-braces==1.14.0